Assignment 2/requirements.txt
numpy
opencv-python
tensorflow
tensorflow-hub
pytube
psycopg2-binary
